+++ dataclass_wizard/serial_json.py
@@ -6,7 +6,7 @@ from .abstractions import AbstractJSONWizard
 from .bases_meta import BaseJSONWizardMeta, LoadMeta, DumpMeta
 from .class_helper import call_meta_initializer_if_needed
 from .dumpers import asdict
-from .loaders_v2 import fromdict, fromlist
+from .v1 import fromdict, fromlist
 # noinspection PyProtectedMember
 from .utils.dataclass_compat import _create_fn, _set_new_attribute
 from .type_def import dataclass_transform
