+++ dataclass_wizard/v1/__init__.py
@@ -1,0 +1,1 @@
+from .loaders import LoadMixin, fromdict, fromlist
